2008-11-11  Esteban Sanchez  <estebans@artica.es>
	* include/js/integria_incident_search.js: Fixed page size in users
	search dialog. User search dialog is a bit bigger now.

	* operation/users/user_search.php: Fixed page size in users search
	dialog.

	* include/js/jquery.tablesorter.pager.js: Fixed a bug that was causing
	double pagination if another search is done.

	* operation/incidents/incident.php: Fixed a bug on the first search,
	which was causing some inconsistent values in the statistics.

	* operation/incidents/incident_detail.php: Fixed a bug that was not
	allowing to add decimal values in the time.

+++ operation/incidents/incident_detail.php
@@ -262,7 +262,7 @@ if ($id) {
 	if ($insert_workunit) {
 		$timestamp = get_parameter ("timestamp");
 		$nota = get_parameter ("nota");
-		$timeused = (int) get_parameter ('duration');
+		$timeused = (float) get_parameter ('duration');
 		$have_cost = (int) get_parameter ('have_cost');
 		$profile = (int) get_parameter ('work_profile');
 		$public = (bool) get_parameter ('public');
